+++ src/wallet.js
@@ -9,6 +9,7 @@ const {
   bnNRlcToBnWei,
   checksummedAddress,
   formatRLC,
+  NULL_BYTES,
 } = require('./utils');
 const foreignBridgeErcToNativeDesc = require('./abi/bridge/ForeignBridgeErcToNative.json');
 const homeBridgeErcToNativeDesc = require('./abi/bridge/HomeBridgeErcToNative.json');
@@ -28,14 +29,14 @@ const ethFaucets = [
   {
     chainName: 'ropsten',
     name: 'faucet.ropsten.be',
-    getETH: address => fetch(`http://faucet.ropsten.be:3001/donate/${address}`)
-      .then(res => res.json())
+    getETH: (address) => fetch(`http://faucet.ropsten.be:3001/donate/${address}`)
+      .then((res) => res.json())
       .catch(() => ({ error: 'ETH faucet is down.' })),
   },
   {
     chainName: 'ropsten',
     name: 'ropsten.faucet.b9lab.com',
-    getETH: address => fetch('https://ropsten.faucet.b9lab.com/tap', {
+    getETH: (address) => fetch('https://ropsten.faucet.b9lab.com/tap', {
       headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
@@ -43,7 +44,7 @@ const ethFaucets = [
       method: 'POST',
       body: JSON.stringify({ toWhom: address }),
     })
-      .then(res => res.json())
+      .then((res) => res.json())
       .catch(() => ({ error: 'ETH faucet is down.' })),
   },
   {
@@ -83,6 +84,63 @@ const getAddress = async (contracts = throwIfMissing()) => {
   return checksummedAddress(address);
 };
 
+const isInWhitelist = async (
+  contracts = throwIfMissing(),
+  address = throwIfMissing(),
+  { strict = true } = {},
+) => {
+  if (contracts.flavour !== 'enterprise') {
+    throw Error('Cannot check authorized eRLC holders on current chain');
+  }
+  const vAddress = await addressSchema({
+    ethProvider: contracts.provider,
+  }).validate(address);
+  try {
+    const eRlcAddress = await wrapCall(contracts.fetchRLCAddress());
+    const eRlcContract = contracts.getRLCContract({ at: eRlcAddress });
+    const isKYC = await wrapCall(eRlcContract.isKYC(vAddress));
+    if (!isKYC && strict) {
+      throw Error(`${vAddress} is not authorized to interact with eRLC`);
+    }
+    return isKYC;
+  } catch (error) {
+    debug('isInWhitelist()', error);
+    throw error;
+  }
+};
+
+const getRlcBalance = async (
+  contracts = throwIfMissing(),
+  address = throwIfMissing(),
+) => {
+  const vAddress = await addressSchema({
+    ethProvider: contracts.provider,
+  }).validate(address);
+  const { isNative } = contracts;
+  if (isNative) {
+    const weiBalance = await contracts.provider.getBalance(vAddress);
+    return truncateBnWeiToBnNRlc(ethersBnToBn(weiBalance));
+  }
+  const rlcAddress = await contracts.fetchRLCAddress();
+  const nRlcBalance = await contracts
+    .getRLCContract({
+      at: rlcAddress,
+    })
+    .balanceOf(vAddress);
+  return ethersBnToBn(nRlcBalance);
+};
+
+const getEthBalance = async (
+  contracts = throwIfMissing(),
+  address = throwIfMissing(),
+) => {
+  const vAddress = await addressSchema({
+    ethProvider: contracts.provider,
+  }).validate(address);
+  const weiBalance = await contracts.provider.getBalance(vAddress);
+  return ethersBnToBn(weiBalance);
+};
+
 const checkBalances = async (
   contracts = throwIfMissing(),
   address = throwIfMissing(),
@@ -91,33 +149,14 @@ const checkBalances = async (
     const vAddress = await addressSchema({
       ethProvider: contracts.provider,
     }).validate(address);
-    const { isNative } = contracts;
-    const getETH = () => contracts.provider.getBalance(vAddress);
-    const balances = {};
-    if (isNative) {
-      const weiBalance = await getETH();
-      Object.assign(balances, {
-        wei: ethersBnToBn(weiBalance),
-        nRLC: truncateBnWeiToBnNRlc(ethersBnToBn(weiBalance)),
-      });
-    } else {
-      const rlcAddress = await contracts.fetchRLCAddress();
-      const getRLC = () => contracts
-        .getRLCContract({
-          at: rlcAddress,
-        })
-        .balanceOf(vAddress)
-        .catch((error) => {
-          debug(error);
-          return 0;
-        });
-
-      const [weiBalance, rlcBalance] = await Promise.all([getETH(), getRLC()]);
-      Object.assign(balances, {
-        wei: ethersBnToBn(weiBalance),
-        nRLC: ethersBnToBn(rlcBalance),
-      });
-    }
+    const [weiBalance, rlcBalance] = await Promise.all([
+      getEthBalance(contracts, vAddress),
+      getRlcBalance(contracts, vAddress),
+    ]);
+    const balances = {
+      wei: weiBalance,
+      nRLC: rlcBalance,
+    };
     debug('balances', balances);
     return balances;
   } catch (error) {
@@ -132,20 +171,16 @@ const getETH = async (
 ) => {
   try {
     const vAddress = await addressSchema().validate(account);
-    const filteredFaucets = ethFaucets.filter(e => e.chainName === chainName);
+    const filteredFaucets = ethFaucets.filter((e) => e.chainName === chainName);
     if (filteredFaucets.length === 0) throw Error(`No ETH faucet on chain ${chainName}`);
     const faucetsResponses = await Promise.all(
-      filteredFaucets.map(faucet => faucet.getETH(vAddress)),
+      filteredFaucets.map((faucet) => faucet.getETH(vAddress)),
     );
     const responses = filteredFaucets.reduce((accu, curr, index) => {
-      accu.push(
-        Object.assign(
-          {
-            name: curr.name,
-          },
-          { response: faucetsResponses[index] },
-        ),
-      );
+      accu.push({
+        name: curr.name,
+        response: faucetsResponses[index],
+      });
       return accu;
     }, []);
     return responses;
@@ -160,7 +195,7 @@ const rlcFaucets = [
     name: 'faucet.iex.ec',
     getRLC: (chainName, address) => fetch(
       `https://api.faucet.iex.ec/getRLC?chainName=${chainName}&address=${address}`,
-    ).then(res => res.json()),
+    ).then((res) => res.json()),
   },
 ];
 
@@ -171,17 +206,13 @@ const getRLC = async (
   try {
     const vAddress = await addressSchema().validate(account);
     const faucetsResponses = await Promise.all(
-      rlcFaucets.map(faucet => faucet.getRLC(chainName, vAddress)),
+      rlcFaucets.map((faucet) => faucet.getRLC(chainName, vAddress)),
     );
     const responses = rlcFaucets.reduce((accu, curr, index) => {
-      accu.push(
-        Object.assign(
-          {
-            name: curr.name,
-          },
-          { response: faucetsResponses[index] },
-        ),
-      );
+      accu.push({
+        name: curr.name,
+        response: faucetsResponses[index],
+      });
       return accu;
     }, []);
     return responses;
@@ -245,16 +276,20 @@ const sendERC20 = async (
 
 const sendETH = async (
   contracts = throwIfMissing(),
-  value = throwIfMissing(),
+  amount = throwIfMissing(),
   to = throwIfMissing(),
 ) => {
   try {
     const vAddress = await addressSchema({
       ethProvider: contracts.provider,
     }).validate(to);
-    const vValue = await weiAmountSchema().validate(value);
+    const vAmount = await weiAmountSchema().validate(amount);
     if (contracts.isNative) throw Error('sendETH() is disabled on sidechain, use sendRLC()');
-    const txHash = await sendNativeToken(contracts, vValue, vAddress);
+    const balance = await getEthBalance(contracts, await getAddress(contracts));
+    if (balance.lt(new BN(vAmount))) {
+      throw Error('Amount to send exceed wallet balance');
+    }
+    const txHash = await sendNativeToken(contracts, vAmount, vAddress);
     return txHash;
   } catch (error) {
     debug('sendETH()', error);
@@ -271,7 +306,17 @@ const sendRLC = async (
     const vAddress = await addressSchema({
       ethProvider: contracts.provider,
     }).validate(to);
+    if (contracts.flavour === 'enterprise') {
+      await isInWhitelist(contracts, await getAddress(contracts), {
+        strict: true,
+      });
+      await isInWhitelist(contracts, to, { strict: true });
+    }
     const vAmount = await nRlcAmountSchema().validate(nRlcAmount);
+    const balance = await getRlcBalance(contracts, await getAddress(contracts));
+    if (balance.lt(new BN(vAmount))) {
+      throw Error('Amount to send exceed wallet balance');
+    }
     if (contracts.isNative) {
       debug('send native token');
       const weiValue = bnNRlcToBnWei(new BN(vAmount)).toString();
@@ -296,6 +341,12 @@ const sweep = async (contracts = throwIfMissing(), to = throwIfMissing()) => {
     const code = await contracts.provider.getCode(vAddressTo);
     if (code !== '0x') {
       throw new Error('Cannot sweep to a contract');
+    }
+    if (contracts.flavour === 'enterprise') {
+      await isInWhitelist(contracts, await getAddress(contracts), {
+        strict: true,
+      });
+      await isInWhitelist(contracts, to, { strict: true });
     }
     let balances = await checkBalances(contracts, userAddress);
     const res = {};
@@ -370,7 +421,10 @@ const bridgeToSidechain = async (
       : undefined;
     const vAmount = await nRlcAmountSchema().validate(nRlcAmount);
     if (contracts.isNative) throw Error('Current chain is a sidechain');
-
+    const balance = await getRlcBalance(contracts, await getAddress(contracts));
+    if (balance.lt(new BN(vAmount))) {
+      throw Error('Amount to bridge exceed wallet balance');
+    }
     const ercBridgeContract = new Contract(
       vBridgeAddress,
       foreignBridgeErcToNativeDesc.abi,
@@ -500,7 +554,7 @@ const bridgeToSidechain = async (
     debug('sendTxHash', sendTxHash);
 
     if (vSidechainBridgeAddress && bridgedContracts) {
-      const waitAffirmationCompleted = txHash => new Promise((resolve) => {
+      const waitAffirmationCompleted = (txHash) => new Promise((resolve) => {
         debug('waitAffirmationCompleted');
         const sidechainBridge = new Contract(
           vSidechainBridgeAddress,
@@ -556,7 +610,10 @@ const bridgeToMainchain = async (
       : undefined;
     const vAmount = await nRlcAmountSchema().validate(nRlcAmount);
     if (!contracts.isNative) throw Error('Current chain is a mainchain');
-
+    const balance = await getRlcBalance(contracts, await getAddress(contracts));
+    if (balance.lt(new BN(vAmount))) {
+      throw Error('Amount to bridge exceed wallet balance');
+    }
     const sidechainBridgeContract = new Contract(
       vBridgeAddress,
       homeBridgeErcToNativeDesc.abi,
@@ -580,7 +637,7 @@ const bridgeToMainchain = async (
       wrapCall(
         sidechainBridgeContract
           .getCurrentDay()
-          .then(currentDay => sidechainBridgeContract.totalSpentPerDay(currentDay)),
+          .then((currentDay) => sidechainBridgeContract.totalSpentPerDay(currentDay)),
       ),
     ]);
     debug('minPerTx', minPerTx.toString());
@@ -621,7 +678,7 @@ const bridgeToMainchain = async (
     debug('sendTxHash', sendTxHash);
 
     if (vMainchainBridgeAddress && bridgedContracts) {
-      const waitRelayedMessage = txHash => new Promise((resolve) => {
+      const waitRelayedMessage = (txHash) => new Promise((resolve) => {
         debug('waitRelayedMessage');
         const mainchainBridge = new Contract(
           vMainchainBridgeAddress,
@@ -658,14 +715,81 @@ const bridgeToMainchain = async (
   }
 };
 
+const wrapEnterpriseRLC = async (
+  contracts = throwIfMissing(),
+  enterpriseContracts,
+  nRlcAmount = throwIfMissing(),
+) => {
+  if (contracts.flavour !== 'standard' || contracts.isNative) {
+    throw Error('Unable to wrap RLC into eRLC on current chain');
+  }
+  if (!enterpriseContracts) {
+    throw Error('Unable to find eRLC on current chain');
+  }
+  const vAmount = await nRlcAmountSchema().validate(nRlcAmount);
+  await isInWhitelist(enterpriseContracts, await getAddress(contracts), {
+    strict: true,
+  });
+  const balance = await getRlcBalance(contracts, await getAddress(contracts));
+  if (balance.lt(new BN(vAmount))) {
+    throw Error('Amount to wrap exceed wallet balance');
+  }
+  try {
+    const eRlcAddress = await wrapCall(enterpriseContracts.fetchRLCAddress());
+    const rlcAddress = await wrapCall(contracts.fetchRLCAddress());
+    const rlcContract = contracts.getRLCContract({ at: rlcAddress });
+    const tx = await wrapSend(
+      rlcContract.approveAndCall(
+        eRlcAddress,
+        vAmount,
+        NULL_BYTES,
+        contracts.txOptions,
+      ),
+    );
+    await wrapWait(tx.wait());
+    return tx.hash;
+  } catch (error) {
+    debug('wrapEnterpriseRLC()', error);
+    throw error;
+  }
+};
+
+const unwrapEnterpriseRLC = async (
+  contracts = throwIfMissing(),
+  nRlcAmount = throwIfMissing(),
+) => {
+  if (contracts.flavour !== 'enterprise' || contracts.isNative) {
+    throw Error('Unable to unwrap eRLC into RLC on current chain');
+  }
+  const vAmount = await nRlcAmountSchema().validate(nRlcAmount);
+  await isInWhitelist(contracts, await getAddress(contracts), { strict: true });
+  const balance = await getRlcBalance(contracts, await getAddress(contracts));
+  if (balance.lt(new BN(vAmount))) {
+    throw Error('Amount to unwrap exceed wallet balance');
+  }
+  try {
+    const eRlcAddress = await wrapCall(contracts.fetchRLCAddress());
+    const eRlcContract = contracts.getRLCContract({ at: eRlcAddress });
+    const tx = await wrapSend(eRlcContract.withdraw(vAmount));
+    await wrapWait(tx.wait());
+    return tx.hash;
+  } catch (error) {
+    debug('unwrapEnterpriseRLC()', error);
+    throw error;
+  }
+};
+
 module.exports = {
   bridgeToMainchain,
   bridgeToSidechain,
   getAddress,
+  isInWhitelist,
   checkBalances,
   getETH,
   getRLC,
   sendETH,
   sendRLC,
   sweep,
+  wrapEnterpriseRLC,
+  unwrapEnterpriseRLC,
 };